+++ tests/pytest/test_controller_documentation.py
@@ -64,49 +64,54 @@ def test_all_controllers_are_documented():
     assert documented_controllers, "No documented controllers were found."
 
 
-PYTEST_MODULE = "shiny.pytest"
-PYTEST_PREFIX = "pytest."
+DOCUMENTED_TESTING_MODULES = {
+    "pytest": "shiny.pytest",
+    "testserver": "shiny.testserver",
+}
+"""Quartodoc prefix -> module, for every testing package the config publishes."""
 
 
-def get_pytest_exports() -> Set[str]:
-    """Return the public names `shiny.pytest` exports via `__all__`."""
-    return set(importlib.import_module(PYTEST_MODULE).__all__)
+def get_module_exports(module: str) -> Set[str]:
+    """Return the public names a module exports via `__all__`."""
+    return set(importlib.import_module(module).__all__)
 
 
-def get_documented_pytest_exports() -> Set[str]:
+def get_documented_names(prefix: str) -> Set[str]:
+    dotted = f"{prefix}."
     return {
-        content[len(PYTEST_PREFIX) :]
+        content[len(dotted) :]
         for section in load_quartodoc_sections(DOCS_CONFIG)
         for content in section_contents(section)
-        if isinstance(content, str) and content.startswith(PYTEST_PREFIX)
+        if isinstance(content, str) and content.startswith(dotted)
     }
 
 
-def test_all_pytest_exports_are_documented():
+@pytest.mark.parametrize("prefix,module", sorted(DOCUMENTED_TESTING_MODULES.items()))
+def test_all_testing_exports_are_documented(prefix: str, module: str):
     """
-    Everything `shiny.pytest` exports must appear in the testing API reference.
+    Everything these modules export must appear in the testing API reference.
 
     `test_all_controllers_are_documented` covers `shiny.playwright.controller`
     only, and `test_quartodoc_configs_have_unique_contents` only rejects
     duplicates, so without this an export could ship with no published docs.
     """
-    exports = get_pytest_exports()
-    documented = get_documented_pytest_exports()
+    exports = get_module_exports(module)
+    documented = get_documented_names(prefix)
 
     error_messages: list[str] = []
     if exports - documented:
         missing_list = "\n".join(
-            sorted(f"  - {PYTEST_PREFIX}{name}" for name in exports - documented)
+            sorted(f"  - {prefix}.{name}" for name in exports - documented)
         )
         error_messages.append(f"Exports missing from {DOCS_CONFIG}:\n{missing_list}")
 
     if documented - exports:
         extra_list = "\n".join(
-            sorted(f"  - {PYTEST_PREFIX}{name}" for name in documented - exports)
+            sorted(f"  - {prefix}.{name}" for name in documented - exports)
         )
         error_messages.append(f"Extraneous entries in {DOCS_CONFIG}:\n{extra_list}")
 
     if error_messages:
         pytest.fail("\n\n".join(error_messages), pytrace=False)
 
-    assert exports, "No `shiny.pytest` exports were found."
+    assert exports, f"No `{module}` exports were found."
